pyproject.toml
[project]
name = "file-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "beautifulsoup4==4.13.3",
  "cachetools==5.5.2",
  "certifi==2025.1.31",
  "cffi==2.0.0",
  "charset-normalizer==3.4.4",
  "cryptography==46.0.3",
  "distro==1.9.0",
  "dotenv==0.9.9",
  "google==3.0.0",
  "google-auth==2.38.0",
  "google-genai==1.10.0",
  "h11==0.14.0",
  "httpcore==1.0.7",
  "httpx==0.28.1",
  "idna==3.10",
  "jiter==0.9.0",
  "openai==1.72.0",
  "pdf2image==1.17.0",
  "pdfminer-six==20250506",
  "pillow==11.1.0",
  "pyasn1==0.6.1",
  "pyasn1-modules==0.4.2",
  "pycparser==2.23",
  "pydantic==2.11.3",
  "pydantic-core==2.33.1",
  "pymupdf==1.25.5",
  "pyside6==6.10.0",
  "pyside6-addons==6.10.0",
  "pyside6-essentials==6.10.0",
  "python-dotenv==1.1.0",
  "requests==2.32.3",
  "rsa==4.9",
  "setuptools==75.8.0",
  "shiboken6==6.10.0",
  "sniffio==1.3.1",
  "soupsieve==2.6",
  "tqdm==4.67.1",
  "torch>=2.0.0",
  "transformers>=4.57.1",
  "typing-extensions==4.13.1",
  "typing-inspection==0.4.0",
  "urllib3==2.3.0",
  "websockets==15.0.1",
  "wheel==0.45.1",
]
